requirements.txt
aiogram>=3.13.0
aiosqlite==0.20.0
asyncpg==0.29.0
apscheduler==3.10.4
python-dotenv==1.0.1
python-dateutil==2.9.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0
aiohttp>=3.9.0
qrcode[pil]==7.4.2
Pillow==10.3.0
phonenumbers==8.13.0
cryptography>=42.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
